Extracted validation logic
